requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.1
django-oauth-toolkit==3.0.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
python-docx==1.1.0
Pillow
reportlab==4.1.0
psycopg2-binary
gunicorn==21.2.0
cryptography==42.0.5
django-ratelimit==4.1.0